requirements.txt
amqp==5.0.9
asgiref==3.2.10
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.1
django-crispy-forms==1.13.0
django-storages==1.12.3
django-tinymce==3.4.0
kombu==5.2.2
Pillow==8.4.0
prompt-toolkit==3.0.24
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
